* ram_size: Type changed from unsigned long into unsigned int.
git-svn-id: svn://svn.code.sf.net/p/colinux/code/branches/devel@1175 688f6386-cf32-0410-9ead-a596f0484e40

+++ src/colinux/common/config.h
@@ -255,11 +255,10 @@ typedef struct co_config {
 	/*
 	 * Size of pseudo physical RAM for this machine (MB).
 	 * 
-	 * The default size is 32MB for systems with more than
-	 * 128MB of physical ram and 16MB for systems with less
-	 * ram.
+	 * The default size is 25% of total RAM for systems with more than
+	 * 128MB of physical ram and 16MB for systems with less ram.
 	 */ 
-	unsigned long ram_size;
+	unsigned int ram_size;
 
 	/*
 	 * The pathname of the initrd file.
